1) Swing compensation table corrected
2) Compiler flags added
3) Disable inactive DDS module during amplitude measurement
4) Amplitude validation correction
5) Extender A ports assignment
6) Reading FPGA versions
7) Debug messages add
8) Amplifier CW and pulse error registers add
9) Acknowledge to PC for any PC to controller write transaction add
10) DDS power on and off function modification
11) DDS SW reset function add
12) Tissue temperature measurement function add
13) Reset all errors function add
14) Add function that resets FPGA after 2.5v power on
15) UART communication reload function add
16) Add UART1 functions for debug
17) Add amplifier parameters coefficients

+++ DDS.h
@@ -20,5 +20,5 @@ int Get_Amplitude_Value(BOOL DDS_Number, BOOL Before_After_Switch);
 
 int Output_Amplitude_Measure(BOOL DDS_Number, BOOL Before_After_Switch);
 
-BOOL Validate_Amplitude_Setting(int Value, BOOL DDS_Number);
+BOOL Validate_Amplitude_Setting(float Value, BOOL DDS_Number);
